pyproject.toml
[tool.poetry]
name = "gb_drf"
version = "0.1.0"
description = ""
authors = ["DashViolin <mymail@srv.org>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "<3.3"
djangorestframework = "^3.13.1"
python-dotenv = "^0.20.0"
# markdown
# django-filter
django-cors-headers = "^3.13.0"
djangorestframework-camel-case = "^1.3.0"

[tool.poetry.dev-dependencies]
isort = "^5.10.1"
black = "^22.3.0"
autoflake = "^1.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
